Changes added as per PR review in EMailAPI

+++ src/main/java/com/amitit/webapp/entity/Enrollment.java
@@ -1,33 +1,29 @@
 package com.amitit.webapp.entity;
 
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import java.time.LocalDate;
+import jakarta.persistence.*;
 
+import lombok.*;
 
 @Entity
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class Enrollment {
-	
+
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int id;
-	
+	private Long id;
+
 	@ManyToOne
 	@JoinColumn(name = "uid")
 	private User user;
-	
+
 	@ManyToOne
 	@JoinColumn(name = "bid")
 	private Batch batch;
-	
+
+//	private LocalDate enrollmentDate;
+//	private String status;
 
 }
